% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.r
\docType{package}
\name{games2-package}
\alias{games2-package}
\title{BR estimation of strategic models}
\description{
An extension to the games package for bias reduced penalized likelihood estimation
}
\section{Acknowledgements}{
 We thank the Wallis Institute of Political
Economy for financial support in making the original games package.
}

\references{
Brenton Kenkel and Curtis S. Signorino.  2014.  "Estimating Extensive Form
Games in R."  \emph{Journal of Statistical Software} 56(8):1--27.

Curtis S. Signorino.  2003.  "Structure and Uncertainty
in Discrete Choice Models."  \emph{Political Analysis} 11:316--344.
}
